+++ osu.Game.Tests/Visual/UserInterface/TestSceneMvisScreen.cs
@@ -1,4 +1,4 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
 using System;
@@ -8,10 +8,6 @@ using osu.Framework.Allocation;
 using osu.Game.Overlays;
 using osu.Game.Screens;
 using osu.Game.Input;
-using osu.Game.Screens.Mvis.UI;
-using osu.Game.Screens.Mvis.Buttons;
-using osu.Game.Screens.Mvis.SideBar;
-using osu.Game.Screens.Mvis;
 using osu.Game.Rulesets.Osu;
 
 namespace osu.Game.Tests.Visual.UserInterface
@@ -19,28 +15,6 @@ namespace osu.Game.Tests.Visual.UserInterface
     [TestFixture]
     public class TestSceneMvisScreen : ScreenTestScene
     {
-        public override IReadOnlyList<Type> RequiredTypes => new[]
-        {
-            typeof(BottomBar),
-            typeof(MvisScreen),
-            typeof(ClockContainer),
-            typeof(BackgroundStoryBoard),
-            typeof(BottomBarButton),
-            typeof(SideBarSettingsPanel),
-            typeof(ToggleableButton),
-            typeof(ToggleableOverlayLockButton),
-            typeof(HoverableProgressBarContainer)
-        };
-
-        private IReadOnlyList<Type> requiredGameDependencies => new[]
-        {
-            typeof(OsuGame),
-            typeof(IdleTracker),
-            typeof(OnScreenDisplay),
-            typeof(NotificationOverlay),
-            typeof(MusicController),
-        };
-
         [Cached]
         private MusicController musicController = new MusicController();