+++ osu.Game.Tests/Visual/UserInterface/TestScenePictureOverlay.cs
@@ -1,10 +1,8 @@
-﻿// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
+// Copyright (c) ppy Pty Ltd <contact@ppy.sh>. Licensed under the MIT Licence.
 // See the LICENCE file in the repository root for full licence text.
 
 using NUnit.Framework;
-using osu.Framework.Graphics.Sprites;
 using osu.Game.Overlays;
-using osu.Game.Overlays.Dialog;
 
 namespace osu.Game.Tests.Visual.UserInterface
 {